requirements.txt
# Core
selenium==4.16.0
undetected-chromedriver==3.5.4

# Google Drive
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Scheduling & Task Management
apscheduler==3.10.4

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Configuration & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0

# Logging & Monitoring
loguru==0.7.2
